Remove GCP RDMA config
The previous environment has the effect of binding all devices to
the first RDMA NIC, reducing bandwidth to 1/8th of the total
capacity. An upstream change will be made to DeepEP or vLLM to
yield a permanent fix, but in the meantime we should revert to
the default behavior which yields about half the target bandwidth.

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>