Enable values override for the Helmfile's inner charts
Currently, if a user wants to deploy a well-lit path following the guides in this repository, they cannot change the values set for the releases in a given Helmfile without changing the values.yaml files in the repo or, in general, with ease.

This commit adds this capability by appending an additional set of
values if defined in the Helm environment at the keys llmdInfra, gatewayInferenceExtension, modelService, respectively for the charts llm-d-infra, gaie-*, ms-*

Users can create a values YAML file, as shown below, and apply it in conjunction with the environment-specific values already set in the available Helmfile templates.

Example usage:
$ helmfile -f /path/to/llmd/repo/guides/inference-scheduling/helmfile.yaml.gotmpl --state-values-file /path/to/additional/values.yaml apply

Example values.yaml:

tolerations: &tolerations
  - effect: NoSchedule
    key: benchmark.llm-d.ai/test-gpu-amd64
    operator: Exists

modelService:
  modelArtifacts:
    uri: "hf://meta-llama/Llama-3.1-8B"
    size: 64Gi
    authSecretName: "llm-d-hf-token"
    name: "meta-llama/Llama-3.1-8B"
  decode:
    tolerations: *tolerations
  prefill:
    tolerations: *tolerations

llmdInfra: {}

gatewayInferenceExtension: {}

# Conflicts:
#	guides/inference-scheduling/helmfile.yaml.gotmpl